Add text annotation examples and refactor existing annotation methods
