+++ examples/working_with_annotations/example_markup_annotation.py
@@ -2,11 +2,65 @@ import sys
 from os import path
 
 import aspose.pdf as ap
+from aspose.pycore import cast
 
 sys.path.append(path.join(path.dirname(__file__), '..'))
 
 from config import set_license, initialize_data_dir
 
+def text_annotation_add(infile, outfile):
+    """Add a text annotation to page 1."""
+    document = ap.Document(infile)
+
+    text_annotation = ap.annotations.TextAnnotation(
+        document.pages[1],
+        ap.Rectangle(299.988, 613.664, 428.708, 680.769, True),
+    )
+    text_annotation.title = "Aspose User"
+    text_annotation.subject = "Sticky Note"
+    text_annotation.contents = "This is a text annotation added by Aspose.PDF for Python via .NET"
+    text_annotation.flags = ap.annotations.AnnotationFlags.PRINT
+    text_annotation.color = ap.Color.blue
+    text_annotation.icon = ap.annotations.TextIcon.HELP
+    popup = ap.annotations.PopupAnnotation(
+        document.pages[1],
+        ap.Rectangle(428.708, 613.664, 528.708, 713.664, True),        
+    )
+    popup.open = True
+
+    text_annotation.popup = popup
+        
+
+    document.pages[1].annotations.add(text_annotation, consider_rotation=False)
+    document.save(outfile)
+
+
+def text_annotation_get(infile, outfile):
+    """Print text annotation rectangles on page 1."""
+    document = ap.Document(infile)
+    text_annotations = [
+        annotation
+        for annotation in document.pages[1].annotations
+        if annotation.annotation_type == ap.annotations.AnnotationType.TEXT
+    ]
+
+    for annotation in text_annotations:
+        print(annotation.rect)
+
+
+def text_annotation_delete(infile, outfile):
+    """Delete text annotations from page 1."""
+    document = ap.Document(infile)
+    text_annotations = [
+        annotation
+        for annotation in document.pages[1].annotations
+        if annotation.annotation_type == ap.annotations.AnnotationType.TEXT
+    ]
+
+    for annotation in text_annotations:
+        document.pages[1].annotations.delete(annotation)
+
+    document.save(outfile)
 
 def caret_annotations_add(infile, outfile):
     """Add a caret annotation to the first page."""
@@ -16,13 +70,16 @@ def caret_annotations_add(infile, outfile):
     # Create Caret Annotation for text insertion
     caret_annotation = ap.annotations.CaretAnnotation(
         page,
-        ap.Rectangle(299.988, 713.664, 308.708, 720.769)
+        ap.Rectangle(299.988, 713.664, 308.708, 720.769, True)
     )
     caret_annotation.title = "Aspose User"
     caret_annotation.subject = "Inserted text 1"
     caret_annotation.flags = ap.annotations.AnnotationFlags.PRINT
     caret_annotation.color = ap.Color.blue
-
+    caret_annotation.popup = ap.annotations.PopupAnnotation(
+        page,
+        ap.Rectangle(310, 713, 410, 730, True)
+    )
     page.annotations.append(caret_annotation)
 
     document.save(outfile)
@@ -84,19 +141,22 @@ def replace_annotations_add(infile, outfile):
     # Create Caret Annotation for text replacement
     caret_annotation = ap.annotations.CaretAnnotation(
         page,
-        ap.Rectangle(361.246, 727.908, 370.081, 735.107)
+        ap.Rectangle(361.246, 727.908, 370.081, 735.107, True)
     )
     caret_annotation.flags = ap.annotations.AnnotationFlags.PRINT
     caret_annotation.subject = "Inserted text 2"
     caret_annotation.title = "Aspose User"
     caret_annotation.color = ap.Color.blue
-
+    caret_annotation.popup = ap.annotations.PopupAnnotation(
+        page,
+        ap.Rectangle(310, 713, 410, 730, True)
+    )
     # Create StrikeOut Annotation
     strikeout_annotation = ap.annotations.StrikeOutAnnotation(
         page,
-        ap.Rectangle(318.407, 727.826, 368.916, 740.098)
+        ap.Rectangle(318.407, 727.826, 368.916, 740.098, True)
     )
-    strikeout_annotation.color = ap.Color.blue
+    strikeout_annotation.color = ap.Color.blue  
     strikeout_annotation.quad_points = [
         ap.Point(321.66, 739.416),
         ap.Point(365.664, 739.416),
@@ -138,9 +198,11 @@ def replace_annotations_get(infile, outfile):
 
     # Iterate through annotations and filter Replace annotations
     for annot in page.annotations:
-        if annot.annotation_type == ap.annotations.AnnotationType.REPLACE:
-            # Print annotation rectangle
-            print(f"Replace annotation rect: {annot.rect}")
+        if annot.annotation_type == ap.annotations.AnnotationType.STRIKE_OUT:
+            sa = cast(ap.annotations.StrikeOutAnnotation, annot)
+            if (sa.in_reply_to is not None and sa.reply_type == ap.annotations.ReplyType.GROUP):
+                # Print annotation rectangle
+                print(f"Replace annotation rect: {sa.rect}")
 
 
 def replace_annotations_delete(infile, outfile):
@@ -166,8 +228,9 @@ def replace_annotations_delete(infile, outfile):
 
     # Collect replace annotations first (avoid modifying collection while iterating)
     replace_annotations = [
-        annot for annot in page.annotations
-        if annot.annotation_type == ap.annotations.AnnotationType.REPLACE
+        cast(ap.annotations.StrikeOutAnnotation, annot)
+        for annot in page.annotations
+        if annot.annotation_type == ap.annotations.AnnotationType.STRIKE_OUT
     ]
 
     # Delete each Replace annotation
@@ -191,16 +254,19 @@ def run_all_examples(data_dir=None, license_path=None):
     input_dir, output_dir = initialize_data_dir(data_dir)
 
     examples = [
-        ("caret_annotations_add", caret_annotations_add),
-        ("caret_annotations_get", caret_annotations_get),
-        ("caret_annotations_delete", caret_annotations_delete),
-        ("replace_annotations_add", replace_annotations_add),
-        ("replace_annotations_get", replace_annotations_get),
-        ("replace_annotations_delete", replace_annotations_delete),
+        ("Add Text Annotation", text_annotation_add),
+        ("Get Text Annotation", text_annotation_get),
+        ("Delete Text Annotation", text_annotation_delete),
+        ("Add Caret Annotation", caret_annotations_add),
+        ("Get Caret Annotation", caret_annotations_get),
+        ("Delete Caret Annotation", caret_annotations_delete),
+        ("Add Replace Annotation", replace_annotations_add),
+        ("Get Replace Annotation", replace_annotations_get),
+        ("Delete Replace Annotation", replace_annotations_delete),
     ]
 
     for name, func in examples:
-        input_file_name = path.join(input_dir, "Annotations.pdf")
+        input_file_name = path.join(input_dir, "sample_n.pdf")
         output_file_name = path.join(output_dir, f"{func.__name__}_out.pdf")
         try:
             func(input_file_name, output_file_name)
